D1035401: Don't create staging queue consumer when consumption target is 0
https://internal.almoctane.com/ui/entity-navigation?p=131002/6001&entityType=work_item&id=1035401

While investigating the above issue, I noticed that sometimes we were processing staging queues when we should not have been. The logs I added in my comment on [I1035401 ](https://internal.almoctane.com/ui/entity-navigation?p=131002/6001&entityType=work_item&id=1035401) show that we are creating a `StagingQueueTargetQueuePair`  even though the staging queue has a consumption target of 0. This results in consuming 1 message from the staging queue and publishing it to a target queue,

This does not happen all the time, only when another staging queue contains > 0 messages for the specific target queue. For example, given this config:

```
"CAF_ADJUST_QUEUE_WEIGHT": "data-enrichment-batch-in»/rorywin1/batch-enrichment-job,0.0",
```

and 1 message on each of the following staging queues:

```
data-enrichment-batch-in»/rorywin1/batch-enrichment-job
data-enrichment-batch-in»/clynch/batch-enrichment-job
```

Then when the Message Distributor runs (default is every 10 seconds), then 1 message will be incorrectly consumed from `data-enrichment-batch-in»/rorywin1/batch-enrichment-job` and published to `data-enrichment-batch-in`.

It happened 15 times in 2 hours when observing the above issue.

This doesn't explain how 1000s of msgs are getting moved from the staging queue in the above issue, but it looks like a bug nonetheless.